requirements.txt
fastapi==0.96.0
keras==2.12.0
numpy==1.23.5
Pillow==9.5.0
Pillow==9.5.0
tensorflow==2.12.0
tensorflow_intel==2.12.0
uvicorn==0.22.0
